Fix aim crash on exit.
git-svn-id: http://miranda.googlecode.com/svn/branches/stable@6341 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/AimOscar/flap.cpp
@@ -1,5 +1,5 @@
 #include "flap.h"
-FLAP::FLAP(char* buf,unsigned short num_bytes)
+FLAP::FLAP(char* buf,int num_bytes)
 {
 	if(FLAP_SIZE>num_bytes)
 	{
